# api-doc.yaml
swagger: '2.0'
info:
  description: 'Webservice documentation for the LCA Collaboration Server'
  version: '1.1.1'
  title: 'LCA Collaboration Server Webservice'
  contact:
    email: 'greve@greendelta.com'
host: 'data.openlca.org'
schemes:
- https
- http
definitions:
...
  Password:
    type: object
    properties:
      password:
        description: 'The new password'
        type: string
        example: 'passw0rd'
      password2:
        description: 'The same password, to ensure password was entered correctly'
        type: string
        example: 'passw0rd'
    required:
      - password
      - password2
  TwoFactorInfo:
    type: object
    properties:
      url:
        description: 'The OtpAuthTotpURL for generating a QR code'
        type: string
        example: 'otpauth://totp/openLCA%20Data%20Server:johndoe?secret=6HWCC6HXXN563FXX&issuer=openLCA+Data+Server'
      key:
        description: 'The two factor authentification secret'
        type: string
        example: '6HWCC6HXXN563FXX'
      enabled:
        description: 'True if two factor authentication is enabled'
        type: boolean
        default: false
    required:
      - url
      - key
...
  UserDescriptor:
    type: object
    properties:
      username:
        description: 'A username, e.g. used for login, etc.'
        type: string
        example: 'johndoe'
      name:
        description: 'A display name for the user'
        type: string
        example: 'John Doe'
    required:
      - username
      - name
  UserSettings:
    type: object
    properties:
      canCreateGroups:
        description: 'True if the user is allowed to create new groups'
        type: boolean
        default: false
      canCreateRepositories:
        description: 'True if the user is allowed to create new repositories'
        type: boolean
        default: false
      noOfRepositories:
        description: '0 for unlimited (if canCreateRepositories is true), otherwise the number of repositories the user is allowed to create'
        type: integer
        format: int32
        default: 0
      maxSize:
        description: '0 for unlimited space (if canCreateRepositories is true), otherwise the maximum size of all user group repositories in bytes'
        type: integer
        format: int64
        default: 0
      messagingEnabled:
        description: 'If true all users can initiate conversations, otherwise admins only'
        type: boolean
        default: true
      messagingRestricted:
        description: 'If true only team members (and admins) can initiate conversations'
        type: boolean
        default: false
      showOnlineStatus:
        description: 'If true the online status is shown in the messanger'
        type: boolean
        default: true
      showReadReceipt:
        description: 'If true read messages will be marked as read in the messanger'
        type: boolean
        default: true
      admin:
        description: 'True if the user is administrator'
        type: boolean
        default: false
      userManager:
        description: 'True if the user is user manager'
        type: boolean
        default: false
      dataManager:
        description: 'True if the user is data manager'
        type: boolean
        default: false
      blockedUsers:
        description: 'Users that are not allowed to initiate conversations'
        type: array
        items:
          $ref: '#/definitions/UserDescriptor'
  User:
    allOf: 
    - $ref: '#/definitions/UserDescriptor'
    properties:
      id: 
        type: string
        example: '15'
      email:
        type: string
        example: 'john.doe@company.com'
      twoFactorAuth:
        description: 'True if two factor authentication is enabled for the user'
        type: boolean
        default: false
      settings:
        $ref: '#/definitions/UserSettings'
    required:
      - id
      - email
...
parameters:
  group:
    name: 'group'
    description: 'The group of the repository'
    in: path
    type: string
    required: true
  repo:
    name: 'repo'
    description: 'The name of the repository'
    in: path
    type: string
    required: true
...
  username:
    name: 'username'
    description: 'The username of the user'
    in: path
    type: string
    required: true
...
  role:
    name: 'role'
    description: 'The role of the user in the repository'
    in: path
    type: string
    required: true
  page:
    name: 'page'
    description: 'page value for pagination, if the value is 0, no paging is applied and all values will be returned'
    in: query
    type: integer
    format: int32
    default: 1
  pageSize:
    name: 'pageSize'
    description: 'Defines the page size for pagination if page is not 0'
    in: query
    type: integer
    format: int32
    default: 10
  filter:
    name: 'filter'
    description: 'Filters results that contain the phrase in their name'
    in: query
    type: string
responses:
  NotLoggedIn:
    description: 'Not authenticated'
...
  NoChangeRolePerm:
    description: 'No permission to CHANGE_ROLE in {group}/{repo}'
  NoChangeRolePermGroup:
    description: 'No permission to CHANGE_ROLE in {group}'
  RepoNotFound:
    description: '{group}/{repo} not found'
  GroupNotFound:
    description: '{group} not found'
  UserNotFound:
    description: '{username} not found'
  UnsupportedSchema:
    description: 'Repository json schema version is not supported'
  NotUserManager:
    description: 'Not a user manager'
paths:
...
  /ws/membership/{group}/user/{username}/{role}:
    post:
      tags:
      - 'Membership (Group)'
      summary: 'Add a user to a group'
      description: 'Add {username} to {group} with {role}. For this operation the client needs permission to CHANGE_ROLE in {group}'
      operationId: 'addUserMembershipForGroup'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/role'
      responses:
        201:
          description: '{username} was added to {group} with {role}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePermGroup'
        404:
          $ref: '#/responses/GroupNotFound'
        409:
          description: '{username} already is a direct member of {group}'
    put:
      tags:
      - 'Membership (Group)'
      summary: 'Update a users role in a group'
      description: 'Update the role of {username} in {group} to {role}. For this operation the client needs permission to CHANGE_ROLE in {group}'
      operationId: 'updateUserMembershipForGroup'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/role'
      responses:
        200:
          description: 'The role of {username} was changed to {role} in {group}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePermGroup'
        404:
          description: '{group} not found or {username} is not a member'
  /ws/membership/{group}/user/{username}:
    delete:
      tags:
      - 'Membership (Group)'
      summary: 'Delete a user from a group'
      description: 'Deletes {username} from {group}. For this operation the client needs permission to CHANGE_ROLE in {group}'
      operationId: 'deleteUserMembershipForGroup'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/username'
      responses:
        200:
          description: '{username} was deletes from {group}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePermGroup'
        404:
          description: '{group} not found or {username} is not a member'
...
  /ws/membership/{group}/{repo}/user/{username}/{role}:
    post:
      tags:
      - 'Membership (Repository)'
      summary: 'Add a user to a repository'
      description: 'Add {username} to {group}/{repo} with {role}. For this operation the client needs permission to CHANGE_ROLE in {group}/{repo}'
      operationId: 'addUserMembershipForRepository'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/role'
      responses:
        201:
          description: '{username} was added to {group}/{repo} with {role}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
        409:
          description: '{username} already is a direct member of {group}/{repo}'
    put:
      tags:
      - 'Membership (Repository)'
      summary: 'Update a users role in a repository'
      description: 'Update the role of {username} in {group}/{repo} to {role}. For this operation the client needs permission to CHANGE_ROLE in {group}/{repo}'
      operationId: 'updateUserMembershipForRepository'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/role'
      responses:
        200:
          description: 'The role of {username} was changed to {role} in {group}/{repo}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePerm'
        404:
          description: '{group}/{repo} not found or {username} is not a member'
        406:
          $ref: '#/responses/UnsupportedSchema'
  /ws/membership/{group}/{repo}/user/{username}:
    delete:
      tags:
      - 'Membership (Repository)'
      summary: 'Delete a user from a repository'
      description: 'Deletes {username} from {group}/{repo}. For this operation the client needs permission to CHANGE_ROLE in {group}/{repo}'
      operationId: 'deleteUserMembershipForRepository'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - $ref: '#/parameters/username'
      responses:
        200:
          description: '{username} was deletes from {group}/{repo}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePerm'
        404:
          description: '{group}/{repo} not found or {username} is not a member'
        406:
          $ref: '#/responses/UnsupportedSchema'
...
  /ws/user:
    get:
      tags: 
      - 'User'
      summary: 'Get users'
      description: 'Loads a list of user descriptors'
      operationId: 'getUsers'
      parameters:
      - $ref: '#/parameters/page'
      - $ref: '#/parameters/pageSize'
      - $ref: '#/parameters/filter'
      - name: 'module'
        description: 'Load a module specific list, MESSAGING loads all users that can be messaged, REVIEW loads all users that have the permission to REVIEW in {repositoryPath}'
        type: string
        enum: ['MESSAGING', 'REVIEW']
        in: query
      - name: 'repositoryPath'
        description: 'The repository id to filter users from (only applies if module=REVIEW)'
        type: string
        in: query
      produces:
      - application/json
      responses:
        200:
          description: 'A list of user descriptors'
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        401:
          $ref: '#/responses/NotLoggedIn'
        404:
          $ref: '#/responses/UserNotFound'
  /ws/user/{username}:
    get:
      tags: 
      - 'User'
      summary: 'Get a user'
      description: 'Loads {username}'
      operationId: 'getUser'
      parameters:
      - $ref: '#/parameters/username'
      produces:
      - application/json
      responses:
        200:
          description: 'If current user is {username}, returns the complete user info, only username and name otherwise'
          schema:
            $ref: '#/definitions/User'
        400:
          description: 'Missing input, invalid username or name or email already in use'    
        401:
          $ref: '#/responses/NotLoggedIn'
        404:
          $ref: '#/responses/UserNotFound'
    put:
      tags: 
      - 'User'
      summary: 'Update a user'
      description: 'Updates {username}'
      operationId: 'updateUser'
      parameters:
      - $ref: '#/parameters/username'
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        200:
          description: 'After successful update, returns the changed user'
          schema:
            $ref: '#/definitions/User'
        400:
          description: 'Missing input or email already in use'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NotUserManager'
        404:
          $ref: '#/responses/UserNotFound'
  /ws/user/avatar/{username}:
    get:
      tags: 
      - 'User'
      summary: 'Get a users avatar'
      description: 'Loads the avatar of {username}'
      operationId: 'getUserAvatar'
      parameters:
      - $ref: '#/parameters/username'
      produces:
      - application/octet-stream
      responses:
        200:
          description: 'Returns the avatar for {group} as byte array'
        401:
          $ref: '#/responses/NotLoggedIn'
        404:
          $ref: '#/responses/UserNotFound'
    put:
      tags: 
      - 'User'
      summary: 'Set user avatar'
      description: 'Sets the avatar of {username}'
      operationId: 'setUserAvatar'
      parameters:
      - $ref: '#/parameters/username'
      consumes:
      - multipart/form-data
      produces:
      - application/octet-stream
      responses:
        200:
          description: 'After successful update, returns the changed avatar'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NotUserManager'
        404:
          $ref: '#/responses/UserNotFound'
  /ws/user/setpassword/{username}:
    put:
      tags: 
      - 'User'
      summary: 'Set password'
      description: 'Sets a new password for {username}'
      operationId: 'setPassword'
      parameters:
      - $ref: '#/parameters/username'
      - name: 'body'
        description: 'The repository id to filter users from (only applies if module=REVIEW)'
        in: body
        schema:
          $ref: '#/definitions/Password'
      consumes:
      - application/json
      responses:
        200:
          description: 'Successfully updated password of {username} to {password}'
        400:
          description: 'Passwords did not match or password was not secure (Password must consist of at least 8 characters and must contain at least 1 digit, 2 different lowercase letters and 2 different uppercase letters)'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NotUserManager'
        404:
          $ref: '#/responses/UserNotFound'
  /ws/user/twoFactorAuth/{username}:
    get:
      tags: 
      - 'User'
      summary: 'Get two factor authentification info'
      description: 'Loads the two factor authentification info of {username}'
      operationId: 'getTwoFactorAuthenticationInfo'
      parameters:
      - $ref: '#/parameters/username'
      produces:
      - application/json
      responses:
        200:
          description: 'Returns the two factor authentification info'
          schema:
            $ref: '#/definitions/TwoFactorInfo'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NotUserManager'
        404:
          $ref: '#/responses/UserNotFound'
  /ws/user/twoFactorAuth/{username}/{enable}:
    put:
      tags: 
      - 'User'
      summary: 'Toggle two factor authentication'
      description: 'Enables/disables configuration of two factor authentification for {username}'
      operationId: 'toggleTwoFactorAuthentication'
      parameters:
      - $ref: '#/parameters/username'
      - name: 'enable'
        description: 'if true, enable two factor authentication, otherwise disable it'
        type: boolean
        in: path
        required: true
      produces:
      - application/json
      responses:
        200:
          description: 'Returns the two factor authentification info if enabled'
          schema:
            $ref: '#/definitions/TwoFactorInfo'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NotUserManager'
        404:
          $ref: '#/responses/UserNotFound'
  /ws/usermanager/user/{username}:
    post:
      tags: 
      - 'User'
      summary: 'Create a new user'
      description: 'Create a new user {username} with generated password. An email containing the password is sent to the email specified for the new user. For this operation the client needs special permission to manage user'
      operationId: 'createUser'
      parameters:
      - $ref: '#/parameters/username'
      - name: 'body'
        description: 'The information of the user to create'
        in: body
        schema:
          $ref: '#/definitions/User'
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        201:
          description: 'User was successfully created'
          schema:
            $ref: '#/definitions/User'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NotUserManager'
        400:
          description: 'Missing input, invalid username or name or email already in use'    
    delete:
      tags: 
      - 'User'
      summary: 'Delete a user'
      description: 'Deletes {username} with generated password. For this operation the client needs special permission to manage user'
      operationId: 'deleteUser'
      parameters:
      - $ref: '#/parameters/username'
      responses:
        200:
          description: 'User was successfully deleted'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NotUserManager'
        404:
          description: '{username} not found'
...